Renamed files to match new component name.
changelog: Upcoming Features,biometric verification,Renamed `DocumentsStep` to `DocumentsAndSelfieStep`
