+++ app/javascript/packages/document-capture/components/documents-and-selfie-step.tsx
